Add files via upload
This is just a basic project on API call showing the profit/loss on any of the Crypto-currency. 
You just have to run main.exe and add your Crypto-currency symbol and number of coins you have bought and in how much you have bought that coin.
Then You just have to click Refresh and there you got the current value of that coin and either you're in Profit or in Loss.